Passing ImageConverterSpec
The ImageConverter spec and ElementInterface were changed to make
testing easier. The `Element::createWriteableElement()` method now
returns an Predmond\HtmlToAmp\Element instance instead of a DOMElement.

The ElementInterface::getNode() method was introduced to get access to
the internal node. This isn't ideal, but is the easiest way to allow an
element to do replacements on the DOMDocument.